+++ ptai-jenkins-plugin/src/main/resources/com/ptsecurity/appsec/ai/ee/utils/ci/integration/plugin/jenkins/reports/Report/config.groovy
@@ -28,7 +28,7 @@ f.entry(
         title: _('locale'),
         field: 'locale') {
     f.select(
-            default: descriptor.getDefaultLocale(),
+            default: descriptor.getDefaultLocale().name(),
     )
 }