+++ ptai-jenkins-plugin/src/main/resources/com/ptsecurity/appsec/ai/ee/utils/ci/integration/plugin/jenkins/serversettings/ServerSettings/config.groovy
@@ -1,5 +1,6 @@
 package com.ptsecurity.appsec.ai.ee.utils.ci.integration.plugin.jenkins.serversettings.ServerSettings
 
+import com.ptsecurity.appsec.ai.ee.utils.ci.integration.Resources
 import lib.CredentialsTagLib
 import lib.FormTagLib
 
@@ -7,7 +8,7 @@ def f = namespace(FormTagLib)
 def c = namespace(CredentialsTagLib)
 
 f.entry(
-        title: _('serverUrl'),
+        title: Resources.i18n_ast_settings_server_url_label(),
         field: 'serverUrl') {
     f.textbox()
 }